+++ src/init_b_a.c
@@ -6,7 +6,7 @@
 /*   By: sfarren <sfarren@student.42malaga.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/18 18:57:16 by sfarren           #+#    #+#             */
-/*   Updated: 2024/10/18 19:57:08 by sfarren          ###   ########.fr       */
+/*   Updated: 2024/10/19 17:51:12 by sfarren          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,11 +16,11 @@ static void	set_target_b(t_stack_node *stack_a, t_stack_node *stack_b)
 {
 	t_stack_node	*current_a;
 	t_stack_node	*target;
-	long			target_index;
+	int				target_index;
 
 	while (stack_b)
 	{
-		target_index = LONG_MAX;
+		target_index = INT_MAX;
 		current_a = stack_a;
 		while (current_a)
 		{
@@ -32,7 +32,7 @@ static void	set_target_b(t_stack_node *stack_a, t_stack_node *stack_b)
 			}
 			current_a = current_a->next;
 		}
-		if (target_index == LONG_MAX)
+		if (target_index == INT_MAX)
 			stack_b->target = find_min(stack_a);
 		else
 			stack_b->target = target;
